move actions into suitable directories
