+++ app/[alias]/(dashboard)/page.tsx
@@ -6,7 +6,7 @@ import {
 import { Suspense } from 'react';
 import { getMembersAction } from '@/app/[alias]/(dashboard)/members/action';
 import { getTransfersOfTokenAction } from '@/app/[alias]/(dashboard)/transfers/actions';
-import { fetchCommunityByAliasAction } from '@/app/_actions/community-actions';
+import { fetchCommunityByAliasAction } from '@/app/[alias]/(dashboard)/_actions/community-actions';
 
 export default async function ProductsPage(props: {
   params: Promise<{ alias: string }>;
