+++ app/[alias]/(dashboard)/transfers/actions.tsx
@@ -2,7 +2,7 @@
 
 import { getServiceRoleClient } from '@/services/chain-db';
 import { getTransfersOfToken } from '@/services/chain-db/transfers';
-import { getAuthUserRoleInCommunityAction } from '@/app/_actions/admin-actions';
+import { getAuthUserRoleInCommunityAction } from '@/app/[alias]/(dashboard)/_actions/admin-actions';
 import { Config } from '@citizenwallet/sdk';
 
 export const getTransfersOfTokenAction = async (args: {
