+++ app/[alias]/(dashboard)/treasury/actions.tsx
@@ -2,7 +2,7 @@
 
 import { getServiceRoleClient } from '@/services/chain-db';
 import { getTreasuryTransfersOfToken } from '@/services/chain-db/transfers';
-import { getAuthUserRoleInCommunityAction } from '@/app/_actions/admin-actions';
+import { getAuthUserRoleInCommunityAction } from '@/app/[alias]/(dashboard)/_actions/admin-actions';
 import { Config } from '@citizenwallet/sdk';
 
 export const getTreasuryTransfersOfTokenAction = async (args: {
